Src/requirements.txt
autogluon
dash==2.18.0
dash_ag_grid==31.3.0
dash_auth
dash_bootstrap_components==1.7.1
diskcache==5.6.3
dowhy
dotenv
flask_login
interpret==0.6.9
kmodes==0.12.2
matplotlib==3.10.0
numpy==1.26.4
pandas==2.2.3
plotly==5.24.1
scipy==1.15.1
seaborn==0.13.2
shap==0.46.0
scikit-learn==1.5.2
statsmodels==0.14.2
torch
umap-learn==0.5.7
xgboost==2.1.3
